[TASK] Adjust creation of new posts to old YYYY/MM/DD scheme
This moves the creation of new blog posts from the command controller
to the Post action controller and emulates the current path scheme of
my website (…blog/posts/2013/02/26/my-blog-post.html).

It's not meant to be the final solution but is necessary for creating
new content on my site at the moment.