[BUGFIX] Prevent PageNotFound exception by paginate widget
If the current page's node has no Posts set at all,
the paginate widget will throw a Page Not Found exception
as if there were no additional posts on some subsequent
page; but this happens also on the initial call.

This patch circumvents this behavior with checking
if there are any posts at all, prior to invoking
the paginate widget.

Fixes #2