[BUGFIX] Content caching fixed for comment creation
The content cache is now built correctly by not using the same entry
identifier for the Page and the Post TS objects and switching the Post
to "cached" instead of (implicit) "embed" mode.

Now the cache entry that is flushed after a new comment is created is
actually used and new comments show up.