FEATURE: Adjust dependencies
Removing the support for Neos version 7 as it is end of life.
Replacing the suggestion of Neos/Swiftmailer to Neos/Symfonymailer as Swiftmailer is also end of life and as we plan to replace Fluid with Fusion AFX, we add the Flowpack/Listable package to be able to have a pagination based on Fusion.