FEATURE: Refactor comment creation form
Instead of a Fluid form with an action controller for comments we use the Neos.Fusion.Form package. As an example we implement a custom from action to save the form data as comment node.

This change also extend the Post node type to be able to enable new comments from the backend.